requirements.txt
numpy
pandas
scipy
scikit-learn
matplotlib
seaborn
plotly
altair
bokeh
pillow
opencv-python
jupyter
pycryptodome
pyrebase4